requirements.txt
flask
python-telegram-bot==13.15
psycopg2-binary
sentence-transformers
spacy
torch
